mimeapps.list: Switch to some XApps by default
This switches the default PDF reader from Evince to XReader, and the default image viewer from Eye of GNOME to XViewer. These applications fit better with the goals and views of Budgie Desktop.

While the file editor Xed is also in the repository, it is almost universally considered to be much inferior to GEdit, or even Mousepad. Since there appears to be no wishes by the GEdit developers to move to GTK4/libadwaita, GEdit is kept as the default text editor.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>